requirements.txt
six==1.12.0
tensorflow==1.13.1
joblib==0.13.2
numpy==1.16.2
h5py==2.9.0
docopt==0.6.2
matplotlib==3.0.3
scikit_learn==0.21.3
PyYAML==5.1.2
